Avoid duplicate urls for the same service on the preview widget
The preview controller checks the request params for duplicate urls
and if some are found, they are removed and a redirect is issued to
the same endpoint. If no duplicates are found, the template is rendered
directly.

Add tests, which needed test ini files in order to load the 'os_preview'
plugin.